split Repository into BaseRepository And NativeRepository
